@@ -1,13 +1,13 @@
 package com.omegafrog.My.piano.app.lesson.entity;
 
-import com.omegafrog.My.piano.app.lesson.dto.UpdateLessonDto;
+import com.omegafrog.My.piano.app.dto.UpdateLessonDto;
 import com.omegafrog.My.piano.app.sheet.entity.Sheet;
 import com.omegafrog.My.piano.app.user.entity.User;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
