@@ -1,12 +1,12 @@
 package com.omegafrog.My.piano.app.sheet.entity;
 
-import com.omegafrog.My.piano.app.sheet.dto.UpdateSheetPostDto;
+import com.omegafrog.My.piano.app.dto.UpdateSheetPostDto;
 import com.omegafrog.My.piano.app.user.entity.User;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
