@@ -1,17 +1,17 @@
 package com.omegafrog.My.piano.app.admin.entity;
 
 
-import com.omegafrog.My.piano.app.admin.dto.AdminDto;
-import com.omegafrog.My.piano.app.admin.dto.UpdateAdminDto;
+import com.omegafrog.My.piano.app.dto.AdminDto;
+import com.omegafrog.My.piano.app.dto.UpdateAdminDto;
 import com.omegafrog.My.piano.app.enums.Position;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 
 @Entity
 @NoArgsConstructor
