@@ -1,22 +1,20 @@
-package com.omegafrog.My.piano.app.admin.dto;
+package com.omegafrog.My.piano.app.dto;
 
 import com.omegafrog.My.piano.app.enums.Position;
 import lombok.Builder;
 import lombok.Data;
 
 @Data
-public class AdminDto {
-    private Long id;
-
+public class UpdateAdminDto {
     private String name;
     private Position position;
     private String email;
 
     @Builder
-    public AdminDto(Long id, String name, Position position, String email) {
-        this.id = id;
+    public UpdateAdminDto(String name, Position position, String email) {
         this.name = name;
         this.position = position;
         this.email = email;
     }
+
 }
