@@ -3,12 +3,12 @@ package com.omegafrog.My.piano.app.sheet.entity;
 import com.omegafrog.My.piano.app.enums.Difficulty;
 import com.omegafrog.My.piano.app.enums.Genre;
 import com.omegafrog.My.piano.app.enums.Instrument;
-import com.omegafrog.My.piano.app.sheet.dto.UpdateSheetDto;
+import com.omegafrog.My.piano.app.dto.UpdateSheetDto;
 import com.omegafrog.My.piano.app.user.entity.User;
+import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
 
 @Entity
 @NoArgsConstructor
