@@ -1,7 +1,7 @@
 package com.omegafrog.My.piano.app.admin.entity;
 
-import com.omegafrog.My.piano.app.admin.dto.AdminDto;
-import com.omegafrog.My.piano.app.admin.dto.UpdateAdminDto;
+import com.omegafrog.My.piano.app.dto.AdminDto;
+import com.omegafrog.My.piano.app.dto.UpdateAdminDto;
 import com.omegafrog.My.piano.app.enums.Position;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -9,8 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @DataJpaTest
 class AdminRepositoryTest {
