@@ -4,12 +4,10 @@ import com.omegafrog.My.piano.app.cart.Cart;
 import com.omegafrog.My.piano.app.enums.Difficulty;
 import com.omegafrog.My.piano.app.enums.Genre;
 import com.omegafrog.My.piano.app.enums.Instrument;
-import com.omegafrog.My.piano.app.sheet.dto.UpdateSheetPostDto;
-import com.omegafrog.My.piano.app.user.dto.UpdateUserDto;
+import com.omegafrog.My.piano.app.dto.UpdateSheetPostDto;
 import com.omegafrog.My.piano.app.user.entity.User;
 import com.omegafrog.My.piano.app.user.vo.LoginMethod;
 import com.omegafrog.My.piano.app.user.vo.PhoneNum;
-import org.apache.juli.logging.Log;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -17,8 +15,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @DataJpaTest
 class SheetPostRepositoryTest {
